Fixes for Rakefile
Firstly invokes were changed to executes. Just using executes in the serve task
was not enough, since lazy_compile calls invoke if changes were made, which
would fail if you needed to compile more than once.

Additionally, File.open is only called after all the output is generated and
stored into a string. Before, if you had an error during CoffeeScript or LESS
compilation, it would not try to recompile on a new request because the output
file's mtime was already updated.